requirements.txt
fastapi==0.116.1
httpx==0.28.1
pydantic==2.11.7
pytest==8.4.1
uvicorn[standard]==0.35.0
redis==5.2.1
SQLAlchemy==2.0.43
asyncpg==0.30.0
PyJWT==2.9.0
passlib==1.7.4
python-dotenv==1.0.1
python-multipart==0.0.9
google-cloud-texttospeech==2.24.0
